feat: Create token_lookup() for tokenizing punctuations
token_lookup generate a dictionary to turn pre-defined punctuation into a token.

This dictionary will be used to tokenize the symbols and add the delimiter (space) around it. This separates each symbols as its own word, making it easier for the neural network to predict the next word. Make sure you don't use a value that could be confused as a word; for example, instead of using the value "dash", try using something like "||dash||".